V1.46
Added 4 types of enemies. Each turret would target their respective color only.

+++ 2DGameEngine/Tools.cpp
@@ -13,10 +13,10 @@ float tools::dy(const SDL_Rect& rect1, const SDL_Rect& rect2) {
 }
 float tools::distance(SDL_Rect rect1, SDL_Rect rect2)
 {
-    float x1 = rect1.x + rect1.w / 2;
-    float y1 = rect1.y + rect1.h / 2;
-    float x2 = rect2.x + rect2.w / 2;
-    float y2 = rect2.y + rect2.h / 2;
+    float x1 = static_cast<float>(rect1.x) + static_cast<float>( rect1.w )/ 2;
+    float y1 = static_cast<float>(rect1.y) + static_cast<float>(rect1.h) / 2;
+    float x2 = static_cast<float>(rect2.x) + static_cast<float>(rect2.w )/ 2;
+    float y2 = static_cast<float>(rect2.y) + static_cast<float>(rect2.h )/ 2;
     float delta;
     delta = sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2));
     return delta;
